Refactor ProfileButton to a Separate Composable File
- Move ProfileButton composable to its own dedicated file for improved code organization and readability.
- Place ProfileButton under a `ui/profile/components` directory, following best practices for sub-composables.
- Update references to ProfileButton in the main screen file to reflect its new location.

This change adheres to the single-responsibility principle, organizing reusable components in a modular structure.
